Fix many warnings
Even a few ones emitted by Borland C++. But these changes do not fix any actual bugs except for literal truncation in fast_utoa for DOS.

The compilation process is now a lot more quiet except with Clang, which complains about shadowing of virtual functions.

+++ examples/tvdemo/ascii.cpp
@@ -98,7 +98,7 @@ void TTable::draw()
 void TTable::charFocused()
 {
     message(owner, evBroadcast, cmAsciiTableCmdBase + cmCharFocused,
-      (void *) (cursor.x + 32 * cursor.y));
+      (void *)(size_t)(cursor.x + 32 * cursor.y));
 }
